address bus done

+++ Components/constants.h
@@ -5,5 +5,5 @@
 #ifndef EMULATOR_CONSTANTS_H
 #define EMULATOR_CONSTANTS_H
 typedef unsigned char Byte;
-typedef unsigned int Word;
+typedef unsigned short int Word;
 #endif //EMULATOR_CONSTANTS_H
